:recycle: refactor: about 페이지 컴포넌트 구조 분리
